+++ src/main/java/ilielucian/demo/bank/config/application/ApplicationConfig.java
@@ -1,7 +1,9 @@
-package ilielucian.demo.bank.application.config;
+package ilielucian.demo.bank.config.application;
 
-import ilielucian.demo.bank.authorization.config.SpringSecurityConfig;
+import ilielucian.demo.bank.config.web.WebAppConfig;
+import ilielucian.demo.bank.config.security.SpringSecurityConfig;
 import ilielucian.demo.bank.bankaccounts.config.BankAccountConfig;
+import ilielucian.demo.bank.config.persistence.HibernateConfig;
 import ilielucian.demo.bank.users.config.UserConfig;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
