+++ src/main/java/ilielucian/demo/bank/users/config/UserConfig.java
@@ -1,6 +1,6 @@
 package ilielucian.demo.bank.users.config;
 
-import ilielucian.demo.bank.authorization.config.SpringSecurityConfig;
+import ilielucian.demo.bank.config.security.SpringSecurityConfig;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
